Initial commit
This commit adds a working copy of the `mgkdbipc` library in the sense
that its test suite passes and is known to be able to deserialise all
KDB types using the 2GB-limited KDB IPC format. It can read decompressed
messages and write KDB IPC.

Further work extending the library is anticiapted.